+++ src/main/java/edu/duke/cs/osprey/markstar/framework/MARKStarBoundAsync.java
@@ -185,11 +185,11 @@ public class MARKStarBoundAsync implements PartitionFunction {
     }
 
     @Override
-    public int getNumConfsEvaluated() {
+    public long getNumConfsEvaluated() {
         return numConfsEnergied;
     }
 
-    public int getNumConfsScored() {
+    public long getNumConfsScored() {
         return numConfsScored;
     }
